Feat/state manager (#32)
* test: mix

* Found bug in Union.

* feat/extract-minipool

* format

* doc: update documentation minipool

* fix(include): change relative to full path

* feat: perfect forwarding

* fix: actual perfect forwarding

* feat: basic clang-tidy

* feat: basic state management for det cea

* fix: expansion working, no memory leak

* feat: add historic_union_list_map to state manager for future evict

* doc: alloc state

* format

* Revert "format"

This reverts commit 2a9ffedc74c930ae5818a414145d5c3494dd150a.

* Revert "doc: alloc state"

This reverts commit 2eea2f81eb6d25ae5790c34090a1247b10aa88bf.

* Revert "feat: add historic_union_list_map to state manager for future evict"

This reverts commit 03c1bbbb24b563eb89ffc517900c355396c54fdd.

* format

* feat: add current iteration to states and evict double linked list

* feat: update evaluator initialization

* The test had segfault because the evaluator was being deallocated after the time-list manager

* feat: add long evaluation test

* fix: rework eviction and format

* fix: disabled swappable due to parsing using it a lot and no change possible

* rename

* feat: update clang tidy files

* fix: unused variables

* feat: add clang-tidy checker

* fix: remove useless clang-tidy files

* feat: add clang-tidy to workflows

* fix: new workflow

* fix: clang-tidy

* fix: clang-tidy

* fix: increase communication delay for test

* fix: add compile commands to workflow for clang-tidy

* fix: add soft link directly

* fix: clang-tidy

* format

* fix: missing tuple import cea.hpp

* fix: missing tuple import logical_cea.hpp

* format

* test: remove repeated test from merge

* fix(bitshift): potential overflow

* fix(spelling)

* fix(state_manager): Move states and map to manager

* feat(state manager): set default starting and max size at 1000

* format

* fix(clang-tidy)

* feat(state_manager): add debug assert for state lists

* fix: remove std::cout

* format

* fix(state_manager): starting/max size

---------

+++ src/core_server/internal/evaluation/predicate_set.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <gmpxx.h>
 
+#include <cstdint>
 #include <iostream>
 
 namespace CORE::Internal::CEA {
@@ -93,9 +94,9 @@ struct PredicateSet {
     std::string out = predicates.get_str(2);
     std::string mask_string = mask.get_str(2);
 
-    int buffer_length = out.size() > mask_string.size()
-                          ? out.size() - mask_string.size()
-                          : mask_string.size() - out.size();
+    uint64_t buffer_length = out.size() > mask_string.size()
+                               ? out.size() - mask_string.size()
+                               : mask_string.size() - out.size();
     std::string buffer = "";
     for (int i = 0; i < buffer_length; i++) {
       buffer += '0';
